Atm_final_version
updated every screen, renamed and overhauled the serial port managing.

updated pinscherm to not automatically put in the pin after entering 4 digits. and now you'll have to click D on the keypad to submit the 

removed reset button on pinscherm

added pressing B and C on the keypad functionality, pressing B removes on digit and pressing C clears the whole textfield.